# #19301: Perform batch allocations when reading device data