--- a/tests/ttnn/benchmark/cpp/CMakeLists.txt
+++ b/tests/ttnn/benchmark/cpp/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(BENCHMARK_SRCS
     "host_tilizer_untilizer/tilizer_untilizer.cpp"
     "padding/pad_rm.cpp"
+    "host_alloc_on_tensor_readback.cpp"
 )
 
 foreach(TEST_SRC ${BENCHMARK_SRCS})